Split WASM channel wrapper responsibilities
Move the WASM channel wrapper's host store, conversion helpers,
credential resolution, and attachment helpers into focused child modules.
Keep the generated WIT bindings and public wrapper API in `mod.rs` so
callers can continue to use `crate::channels::wasm::wrapper::*`.
