+++ src/channels/wasm/wrapper/mod.rs
@@ -37,7 +37,6 @@ use tokio_stream::wrappers::ReceiverStream;
 use uuid::Uuid;
 use wasmtime::Store;
 use wasmtime::component::{HasSelf, Linker};
-use wasmtime_wasi::{ResourceTable, WasiCtx, WasiCtxBuilder, WasiCtxView, WasiView};
 
 use crate::channels::wasm::capabilities::ChannelCapabilities;
 use crate::channels::wasm::error::WasmChannelError;
@@ -53,13 +52,24 @@ use crate::channels::{
 use crate::db::{SettingKey, UserId};
 use crate::error::ChannelError;
 use crate::pairing::PairingStore;
-use crate::safety::LeakDetector;
 use crate::secrets::SecretsStore;
-use crate::tools::wasm::LogLevel;
-use crate::tools::wasm::WasmResourceLimiter;
-use crate::tools::wasm::credential_injector::{
-    InjectedCredentials, host_matches_pattern, inject_credential,
+
+mod attachments;
+mod convert;
+mod credentials;
+mod store;
+
+use attachments::read_attachments;
+use convert::{
+    clone_wit_status_update, convert_channel_config, convert_http_response, status_to_wit,
+    wit_channel,
 };
+use credentials::resolve_channel_host_credentials;
+use store::{ChannelStoreData, ResolvedHostCredential};
+
+#[cfg(test)]
+use attachments::mime_from_extension;
+pub use convert::HttpResponse;
 
 // Generate component model bindings from the WIT file
 wasmtime::component::bindgen!({
@@ -69,580 +79,6 @@ wasmtime::component::bindgen!({
         // Use our own store data type
     },
 });
-
-/// Pre-resolved credential for host-based injection.
-///
-/// Built before each WASM execution by decrypting secrets from the store.
-/// Applied per-request by matching the URL host against `host_patterns`.
-/// WASM channels never see the raw secret values.
-#[derive(Clone)]
-struct ResolvedHostCredential {
-    /// Host patterns this credential applies to (e.g., "api.slack.com").
-    host_patterns: Vec<String>,
-    /// Headers to add to matching requests (e.g., "Authorization: Bearer ...").
-    headers: HashMap<String, String>,
-    /// Query parameters to add to matching requests.
-    query_params: HashMap<String, String>,
-    /// Raw secret value for redaction in error messages.
-    secret_value: String,
-}
-
-/// Store data for WASM channel execution.
-///
-/// Contains the resource limiter, channel-specific host state, and WASI context.
-struct ChannelStoreData {
-    limiter: WasmResourceLimiter,
-    host_state: ChannelHostState,
-    wasi: WasiCtx,
-    table: ResourceTable,
-    /// Injected credentials for URL substitution (e.g., bot tokens).
-    /// Keys are placeholder names like "TELEGRAM_BOT_TOKEN".
-    credentials: HashMap<String, String>,
-    /// Pre-resolved credentials for automatic host-based injection.
-    /// Applied per-request by matching the URL host against host_patterns.
-    host_credentials: Vec<ResolvedHostCredential>,
-    /// Pairing store for DM pairing (guest access control).
-    pairing_store: Arc<PairingStore>,
-    /// Dedicated tokio runtime for HTTP requests, lazily initialized.
-    /// Reused across multiple `http_request` calls within one execution.
-    http_runtime: Option<tokio::runtime::Runtime>,
-}
-
-impl ChannelStoreData {
-    fn new(
-        memory_limit: u64,
-        channel_name: &str,
-        capabilities: ChannelCapabilities,
-        credentials: HashMap<String, String>,
-        host_credentials: Vec<ResolvedHostCredential>,
-        pairing_store: Arc<PairingStore>,
-    ) -> Self {
-        // Create a minimal WASI context (no filesystem, no env vars for security)
-        let wasi = WasiCtxBuilder::new().build();
-
-        Self {
-            limiter: WasmResourceLimiter::new(memory_limit),
-            host_state: ChannelHostState::new(channel_name, capabilities),
-            wasi,
-            table: ResourceTable::new(),
-            credentials,
-            host_credentials,
-            pairing_store,
-            http_runtime: None,
-        }
-    }
-
-    /// Inject credentials into a string by replacing placeholders.
-    ///
-    /// Replaces patterns like `{TELEGRAM_BOT_TOKEN}` or `{WHATSAPP_ACCESS_TOKEN}`
-    /// with actual values from the injected credentials map. This allows WASM
-    /// channels to reference credentials without ever seeing the actual values.
-    ///
-    /// Works on URLs, headers, or any string with credential placeholders.
-    fn inject_credentials(&self, input: &str, context: &str) -> String {
-        let mut result = input.to_string();
-
-        tracing::debug!(
-            input_preview = %input.chars().take(100).collect::<String>(),
-            context = %context,
-            credential_count = self.credentials.len(),
-            credential_names = ?self.credentials.keys().collect::<Vec<_>>(),
-            "Injecting credentials"
-        );
-
-        // Replace all known placeholders from the credentials map
-        for (name, value) in &self.credentials {
-            let placeholder = format!("{{{}}}", name);
-            if result.contains(&placeholder) {
-                tracing::debug!(
-                    placeholder = %placeholder,
-                    context = %context,
-                    "Found and replacing credential placeholder"
-                );
-                result = result.replace(&placeholder, value);
-            }
-        }
-
-        // Check if any placeholders remain (indicates missing credential)
-        if result.contains('{') && result.contains('}') {
-            // Only warn if it looks like an unresolved placeholder (not JSON braces)
-            let brace_pattern = regex::Regex::new(r"\{[A-Z_]+\}").ok();
-            if let Some(re) = brace_pattern
-                && re.is_match(&result)
-            {
-                tracing::warn!(
-                    context = %context,
-                    "String may contain unresolved credential placeholders"
-                );
-            }
-        }
-
-        result
-    }
-
-    /// Replace injected credential values with `[REDACTED]` in text.
-    ///
-    /// Prevents credentials from leaking through error messages, logs, or
-    /// return values to WASM. reqwest::Error includes the full URL in its
-    /// Display output, so any error from an injected-URL request will
-    /// contain the raw credential unless we scrub it.
-    ///
-    /// Scrubs raw, URL-encoded, and Base64-encoded forms of each secret
-    /// to prevent exfiltration via encoded representations in error strings.
-    fn redact_credentials(&self, text: &str) -> String {
-        let mut result = text.to_string();
-        for (name, value) in &self.credentials {
-            if !value.is_empty() {
-                let tag = format!("[REDACTED:{}]", name);
-                result = result.replace(value, &tag);
-                // Also redact URL-encoded form (covers secrets in query strings)
-                let encoded = urlencoding::encode(value);
-                if encoded != *value {
-                    result = result.replace(encoded.as_ref(), &tag);
-                }
-            }
-        }
-        for cred in &self.host_credentials {
-            if !cred.secret_value.is_empty() {
-                let tag = "[REDACTED:host_credential]";
-                result = result.replace(&cred.secret_value, tag);
-                // Also redact URL-encoded form (covers secrets injected as query params)
-                let encoded = urlencoding::encode(&cred.secret_value);
-                if encoded.as_ref() != cred.secret_value {
-                    result = result.replace(encoded.as_ref(), tag);
-                }
-            }
-        }
-        result
-    }
-
-    /// Inject pre-resolved host credentials into the request.
-    ///
-    /// Matches the URL host against each resolved credential's host_patterns.
-    /// Matching credentials have their headers merged and query params appended.
-    fn inject_host_credentials(
-        &self,
-        url_host: &str,
-        headers: &mut HashMap<String, String>,
-        url: &mut String,
-    ) {
-        for cred in &self.host_credentials {
-            let matches = cred
-                .host_patterns
-                .iter()
-                .any(|pattern| host_matches_pattern(url_host, pattern));
-
-            if !matches {
-                continue;
-            }
-
-            // Merge injected headers (host credentials take precedence)
-            for (key, value) in &cred.headers {
-                headers.insert(key.clone(), value.clone());
-            }
-
-            // Append query parameters to URL
-            if !cred.query_params.is_empty() {
-                if let Ok(mut parsed_url) = url::Url::parse(url) {
-                    for (name, value) in &cred.query_params {
-                        parsed_url.query_pairs_mut().append_pair(name, value);
-                    }
-                    *url = parsed_url.to_string();
-                } else {
-                    tracing::warn!(url = %url, "Could not parse URL to inject query parameters; skipping injection");
-                }
-            }
-        }
-    }
-}
-
-// Implement WasiView to provide WASI context and resource table
-impl WasiView for ChannelStoreData {
-    fn ctx(&mut self) -> WasiCtxView<'_> {
-        WasiCtxView {
-            ctx: &mut self.wasi,
-            table: &mut self.table,
-        }
-    }
-}
-
-// Implement the generated Host trait for channel-host interface
-impl near::agent::channel_host::Host for ChannelStoreData {
-    fn log(&mut self, level: near::agent::channel_host::LogLevel, message: String) {
-        let log_level = match level {
-            near::agent::channel_host::LogLevel::Trace => LogLevel::Trace,
-            near::agent::channel_host::LogLevel::Debug => LogLevel::Debug,
-            near::agent::channel_host::LogLevel::Info => LogLevel::Info,
-            near::agent::channel_host::LogLevel::Warn => LogLevel::Warn,
-            near::agent::channel_host::LogLevel::Error => LogLevel::Error,
-        };
-        let _ = self.host_state.log(log_level, message);
-    }
-
-    fn now_millis(&mut self) -> u64 {
-        self.host_state.now_millis()
-    }
-
-    fn workspace_read(&mut self, path: String) -> Option<String> {
-        self.host_state.workspace_read(&path).ok().flatten()
-    }
-
-    fn workspace_write(&mut self, path: String, content: String) -> Result<(), String> {
-        self.host_state
-            .workspace_write(&path, content)
-            .map_err(|e| e.to_string())
-    }
-
-    fn http_request(
-        &mut self,
-        method: String,
-        url: String,
-        headers_json: String,
-        body: Option<Vec<u8>>,
-        timeout_ms: Option<u32>,
-    ) -> Result<near::agent::channel_host::HttpResponse, String> {
-        tracing::info!(
-            method = %method,
-            original_url = %url,
-            body_len = body.as_ref().map(|b| b.len()).unwrap_or(0),
-            "WASM http_request called"
-        );
-
-        // Preserve the raw request for leak scanning before any host-side
-        // placeholder resolution. WASM only sees placeholders, not real secrets.
-        let raw_url = url.clone();
-        let injected_url = self.inject_credentials(&raw_url, "url");
-
-        // Log whether injection happened (without revealing the token)
-        let url_changed = injected_url != url;
-        tracing::info!(url_changed = url_changed, "URL after credential injection");
-
-        // Check if HTTP is allowed for this URL
-        self.host_state
-            .check_http_allowed(&injected_url, &method)
-            .map_err(|e| {
-                tracing::error!(error = %e, "HTTP not allowed");
-                format!("HTTP not allowed: {}", e)
-            })?;
-
-        // Record the request for rate limiting
-        self.host_state.record_http_request().map_err(|e| {
-            tracing::error!(error = %e, "Rate limit exceeded");
-            format!("Rate limit exceeded: {}", e)
-        })?;
-
-        // Parse the raw header values supplied by WASM.
-        let raw_headers: std::collections::HashMap<String, String> =
-            serde_json::from_str(&headers_json).unwrap_or_default();
-
-        // Leak scan runs on WASM-provided values before any host-side
-        // credential injection. This prevents false positives where the
-        // resolved secret value would otherwise be attributed to the channel.
-        let leak_detector = LeakDetector::new();
-        let raw_header_vec: Vec<(String, String)> = raw_headers
-            .iter()
-            .map(|(k, v)| (k.clone(), v.clone()))
-            .collect();
-
-        leak_detector
-            .scan_http_request(&raw_url, &raw_header_vec, body.as_deref())
-            .map_err(|e| format!("Potential secret leak blocked: {}", e))?;
-
-        let mut headers: std::collections::HashMap<String, String> = raw_headers
-            .into_iter()
-            .map(|(k, v)| {
-                (
-                    k.clone(),
-                    self.inject_credentials(&v, &format!("header:{}", k)),
-                )
-            })
-            .collect();
-
-        let headers_changed = headers
-            .values()
-            .any(|v| v.contains("Bearer ") && !v.contains('{'));
-        tracing::debug!(
-            header_count = headers.len(),
-            headers_changed = headers_changed,
-            "Parsed and injected request headers"
-        );
-
-        let mut url = injected_url;
-
-        // Inject pre-resolved host credentials (Bearer tokens, API keys, etc.)
-        // after the leak scan so host-injected secrets don't trigger false positives.
-        if let Some(host) = extract_host_from_url(&url) {
-            self.inject_host_credentials(&host, &mut headers, &mut url);
-        }
-
-        // Get the max response size from capabilities (default 10MB).
-        let max_response_bytes = self
-            .host_state
-            .capabilities()
-            .tool_capabilities
-            .http
-            .as_ref()
-            .map(|h| h.max_response_bytes)
-            .unwrap_or(10 * 1024 * 1024);
-
-        // Make the HTTP request using a dedicated single-threaded runtime.
-        // We're inside spawn_blocking, so we can't rely on the main runtime's
-        // I/O driver (it may be busy with WASM compilation or other startup work).
-        // A dedicated runtime gives us our own I/O driver and avoids contention.
-        // The runtime is lazily created and reused across calls within one execution.
-        if self.http_runtime.is_none() {
-            self.http_runtime = Some(
-                tokio::runtime::Builder::new_current_thread()
-                    .enable_all()
-                    .build()
-                    .map_err(|e| format!("Failed to create HTTP runtime: {e}"))?,
-            );
-        }
-        let rt = self.http_runtime.as_ref().expect("just initialized");
-        let result = rt.block_on(async {
-            let client = reqwest::Client::builder()
-                .connect_timeout(std::time::Duration::from_secs(10))
-                .build()
-                .map_err(|e| format!("Failed to build HTTP client: {e}"))?;
-
-            let mut request = match method.to_uppercase().as_str() {
-                "GET" => client.get(&url),
-                "POST" => client.post(&url),
-                "PUT" => client.put(&url),
-                "DELETE" => client.delete(&url),
-                "PATCH" => client.patch(&url),
-                "HEAD" => client.head(&url),
-                _ => return Err(format!("Unsupported HTTP method: {}", method)),
-            };
-
-            // Add headers
-            for (key, value) in headers {
-                request = request.header(&key, &value);
-            }
-
-            // Add body if present
-            if let Some(body_bytes) = body {
-                request = request.body(body_bytes);
-            }
-
-            // Send request with caller-specified timeout (default 30s, max 5min).
-            let timeout_ms = timeout_ms.unwrap_or(30_000).min(300_000) as u64;
-            let timeout = std::time::Duration::from_millis(timeout_ms);
-            let response = request.timeout(timeout).send().await.map_err(|e| {
-                // Walk the full error chain so we get the actual root cause
-                // (DNS, TLS, connection refused, etc.) instead of just
-                // "error sending request for url (...)".
-                let mut chain = format!("HTTP request failed: {}", e);
-                let mut source = std::error::Error::source(&e);
-                while let Some(cause) = source {
-                    chain.push_str(&format!(" -> {}", cause));
-                    source = cause.source();
-                }
-                chain
-            })?;
-
-            let status = response.status().as_u16();
-            let response_headers: std::collections::HashMap<String, String> = response
-                .headers()
-                .iter()
-                .filter_map(|(k, v)| {
-                    v.to_str()
-                        .ok()
-                        .map(|v| (k.as_str().to_string(), v.to_string()))
-                })
-                .collect();
-            let headers_json = serde_json::to_string(&response_headers).unwrap_or_default();
-
-            // Enforce max response body size to prevent memory exhaustion.
-            let max_response = max_response_bytes;
-            if let Some(cl) = response.content_length()
-                && cl as usize > max_response
-            {
-                return Err(format!(
-                    "Response body too large: {} bytes exceeds limit of {} bytes",
-                    cl, max_response
-                ));
-            }
-            let body = response
-                .bytes()
-                .await
-                .map_err(|e| format!("Failed to read response body: {}", e))?;
-            if body.len() > max_response {
-                return Err(format!(
-                    "Response body too large: {} bytes exceeds limit of {} bytes",
-                    body.len(),
-                    max_response
-                ));
-            }
-            let body = body.to_vec();
-
-            tracing::info!(
-                status = status,
-                body_len = body.len(),
-                "HTTP response received"
-            );
-
-            // Log response body for debugging (truncated at char boundary)
-            if let Ok(body_str) = std::str::from_utf8(&body) {
-                let truncated = if body_str.chars().count() > 500 {
-                    format!("{}...", body_str.chars().take(500).collect::<String>())
-                } else {
-                    body_str.to_string()
-                };
-                tracing::debug!(body = %truncated, "Response body");
-            }
-
-            // Leak detection on response body (best-effort)
-            if let Ok(body_str) = std::str::from_utf8(&body) {
-                leak_detector
-                    .scan_and_clean(body_str)
-                    .map_err(|e| format!("Potential secret leak in response: {}", e))?;
-            }
-
-            Ok(near::agent::channel_host::HttpResponse {
-                status,
-                headers_json,
-                body,
-            })
-        });
-
-        // Scrub credential values from error messages before logging or returning
-        // to WASM. reqwest::Error includes the full URL (with injected credentials)
-        // in its Display output.
-        let result = result.map_err(|e| self.redact_credentials(&e));
-
-        match &result {
-            Ok(resp) => {
-                tracing::info!(status = resp.status, "http_request completed successfully");
-            }
-            Err(e) => {
-                tracing::error!(error = %e, "http_request failed");
-            }
-        }
-
-        result
-    }
-
-    fn secret_exists(&mut self, name: String) -> bool {
-        self.host_state.secret_exists(&name)
-    }
-
-    fn emit_message(&mut self, msg: near::agent::channel_host::EmittedMessage) {
-        tracing::info!(
-            user_id = %msg.user_id,
-            user_name = ?msg.user_name,
-            content_len = msg.content.len(),
-            attachment_count = msg.attachments.len(),
-            "WASM emit_message called"
-        );
-
-        let attachments: Vec<crate::channels::wasm::host::Attachment> = msg
-            .attachments
-            .into_iter()
-            .map(|a| {
-                // Parse extras-json for well-known fields
-                let extras: serde_json::Value = if a.extras_json.is_empty() {
-                    serde_json::Value::Null
-                } else {
-                    serde_json::from_str(&a.extras_json).unwrap_or(serde_json::Value::Null)
-                };
-                let duration_secs = extras
-                    .get("duration_secs")
-                    .and_then(|v| v.as_u64())
-                    .map(|v| v as u32);
-
-                // Merge stored binary data (from store-attachment-data host call)
-                let data = self
-                    .host_state
-                    .remove_attachment_data(&a.id)
-                    .unwrap_or_default();
-
-                crate::channels::wasm::host::Attachment {
-                    id: a.id,
-                    mime_type: a.mime_type,
-                    filename: a.filename,
-                    size_bytes: a.size_bytes,
-                    source_url: a.source_url,
-                    storage_key: a.storage_key,
-                    extracted_text: a.extracted_text,
-                    data,
-                    duration_secs,
-                }
-            })
-            .collect();
-
-        let mut emitted = EmittedMessage::new(msg.user_id.clone(), msg.content.clone());
-        if let Some(name) = msg.user_name {
-            emitted = emitted.with_user_name(name);
-        }
-        if let Some(tid) = msg.thread_id {
-            emitted = emitted.with_thread_id(tid);
-        }
-        emitted = emitted.with_metadata(msg.metadata_json);
-        emitted = emitted.with_attachments(attachments);
-
-        match self.host_state.emit_message(emitted) {
-            Ok(()) => {
-                tracing::info!("Message emitted to host state successfully");
-            }
-            Err(e) => {
-                tracing::error!(error = %e, "Failed to emit message to host state");
-            }
-        }
-    }
-
-    fn store_attachment_data(
-        &mut self,
-        attachment_id: String,
-        data: Vec<u8>,
-    ) -> Result<(), String> {
-        tracing::debug!(
-            attachment_id = %attachment_id,
-            size = data.len(),
-            "WASM store_attachment_data called"
-        );
-        self.host_state
-            .store_attachment_data(&attachment_id, data)
-            .map_err(|e| e.to_string())
-    }
-
-    fn pairing_upsert_request(
-        &mut self,
-        channel: String,
-        id: String,
-        meta_json: String,
-    ) -> Result<near::agent::channel_host::PairingUpsertResult, String> {
-        let meta = if meta_json.is_empty() {
-            None
-        } else {
-            serde_json::from_str(&meta_json).ok()
-        };
-        match self.pairing_store.upsert_request(&channel, &id, meta) {
-            Ok(r) => Ok(near::agent::channel_host::PairingUpsertResult {
-                code: r.code,
-                created: r.created,
-            }),
-            Err(e) => Err(e.to_string()),
-        }
-    }
-
-    fn pairing_is_allowed(
-        &mut self,
-        channel: String,
-        id: String,
-        username: Option<String>,
-    ) -> Result<bool, String> {
-        self.pairing_store
-            .is_sender_allowed(&channel, &id, username.as_deref())
-            .map_err(|e| e.to_string())
-    }
-
-    fn pairing_read_allow_from(&mut self, channel: String) -> Result<Vec<String>, String> {
-        self.pairing_store
-            .read_allow_from(&channel)
-            .map_err(|e| e.to_string())
-    }
-}
 
 /// A WASM-based channel implementing the Channel trait.
 #[allow(dead_code)]
@@ -2644,412 +2080,6 @@ impl NativeChannel for SharedWasmChannel {
     async fn shutdown(&self) -> Result<(), ChannelError> {
         self.inner.shutdown().await
     }
-}
-
-// ============================================================================
-// WIT Type Conversion Helpers
-// ============================================================================
-
-// Type aliases for the generated WIT types (exported interface)
-use exports::near::agent::channel as wit_channel;
-
-/// Convert WIT-generated ChannelConfig to our internal type.
-fn convert_channel_config(wit: wit_channel::ChannelConfig) -> ChannelConfig {
-    ChannelConfig {
-        display_name: wit.display_name,
-        http_endpoints: wit
-            .http_endpoints
-            .into_iter()
-            .map(
-                |ep| crate::channels::wasm::schema::HttpEndpointConfigSchema {
-                    path: ep.path,
-                    methods: ep.methods,
-                    require_secret: ep.require_secret,
-                },
-            )
-            .collect(),
-        poll: wit
-            .poll
-            .map(|p| crate::channels::wasm::schema::PollConfigSchema {
-                interval_ms: p.interval_ms,
-                enabled: p.enabled,
-            }),
-    }
-}
-
-/// Convert WIT-generated OutgoingHttpResponse to our HttpResponse type.
-fn convert_http_response(wit: wit_channel::OutgoingHttpResponse) -> HttpResponse {
-    let headers = serde_json::from_str(&wit.headers_json).unwrap_or_default();
-    HttpResponse {
-        status: wit.status,
-        headers,
-        body: wit.body,
-    }
-}
-
-/// Convert a StatusUpdate + metadata into the WIT StatusUpdate type.
-fn truncate_status_text(input: &str, max_chars: usize) -> String {
-    let mut iter = input.chars();
-    let truncated: String = iter.by_ref().take(max_chars).collect();
-    if iter.next().is_some() {
-        format!("{}...", truncated)
-    } else {
-        truncated
-    }
-}
-
-fn status_to_wit(status: &StatusUpdate, metadata: &serde_json::Value) -> wit_channel::StatusUpdate {
-    let metadata_json = serde_json::to_string(metadata).unwrap_or_default();
-
-    match status {
-        StatusUpdate::Thinking(msg) => wit_channel::StatusUpdate {
-            status: wit_channel::StatusType::Thinking,
-            message: msg.clone(),
-            metadata_json,
-        },
-        StatusUpdate::ToolStarted { name } => wit_channel::StatusUpdate {
-            status: wit_channel::StatusType::ToolStarted,
-            message: format!("Tool started: {}", name),
-            metadata_json,
-        },
-        StatusUpdate::ToolCompleted { name, success, .. } => wit_channel::StatusUpdate {
-            status: wit_channel::StatusType::ToolCompleted,
-            message: format!(
-                "Tool completed: {} ({})",
-                name,
-                if *success { "ok" } else { "failed" }
-            ),
-            metadata_json,
-        },
-        StatusUpdate::ToolResult { name, preview } => wit_channel::StatusUpdate {
-            status: wit_channel::StatusType::ToolResult,
-            message: format!(
-                "Tool result: {}\n{}",
-                name,
-                truncate_status_text(preview, 280)
-            ),
-            metadata_json,
-        },
-        StatusUpdate::StreamChunk(chunk) => wit_channel::StatusUpdate {
-            status: wit_channel::StatusType::Thinking,
-            message: chunk.clone(),
-            metadata_json,
-        },
-        StatusUpdate::Status(msg) => {
-            // Map well-known status strings to WIT types (case-insensitive
-            // to stay consistent with is_terminal_text_status and the
-            // Telegram-side classify_status_update).
-            let trimmed = msg.trim();
-            let status_type = if trimmed.eq_ignore_ascii_case("done") {
-                wit_channel::StatusType::Done
-            } else if trimmed.eq_ignore_ascii_case("interrupted") {
-                wit_channel::StatusType::Interrupted
-            } else {
-                wit_channel::StatusType::Status
-            };
-            wit_channel::StatusUpdate {
-                status: status_type,
-                message: msg.clone(),
-                metadata_json,
-            }
-        }
-        StatusUpdate::ApprovalNeeded {
-            request_id,
-            tool_name,
-            description,
-            ..
-        } => wit_channel::StatusUpdate {
-            status: wit_channel::StatusType::ApprovalNeeded,
-            message: format!(
-                "Approval needed for tool '{}'. {}\nRequest ID: {}\nReply with: yes (or /approve), no (or /deny), or always (or /always).",
-                tool_name, description, request_id
-            ),
-            metadata_json,
-        },
-        StatusUpdate::JobStarted {
-            job_id,
-            title,
-            browse_url,
-        } => wit_channel::StatusUpdate {
-            status: wit_channel::StatusType::JobStarted,
-            message: format!("Job started: {} ({})\n{}", title, job_id, browse_url),
-            metadata_json,
-        },
-        StatusUpdate::AuthRequired {
-            extension_name,
-            instructions,
-            auth_url,
-            setup_url,
-        } => wit_channel::StatusUpdate {
-            status: wit_channel::StatusType::AuthRequired,
-            message: {
-                let mut lines = vec![format!("Authentication required for {}.", extension_name)];
-                if let Some(text) = instructions
-                    && !text.trim().is_empty()
-                {
-                    lines.push(text.trim().to_string());
-                }
-                if let Some(url) = auth_url {
-                    lines.push(format!("Auth URL: {}", url));
-                }
-                if let Some(url) = setup_url {
-                    lines.push(format!("Setup URL: {}", url));
-                }
-                lines.join("\n")
-            },
-            metadata_json,
-        },
-        StatusUpdate::AuthCompleted {
-            extension_name,
-            success,
-            message,
-        } => wit_channel::StatusUpdate {
-            status: wit_channel::StatusType::AuthCompleted,
-            message: format!(
-                "Authentication {} for {}. {}",
-                if *success { "completed" } else { "failed" },
-                extension_name,
-                message
-            ),
-            metadata_json,
-        },
-        StatusUpdate::ImageGenerated { path, .. } => wit_channel::StatusUpdate {
-            status: wit_channel::StatusType::Status,
-            message: match path {
-                Some(p) => format!("[image] {}", p),
-                None => "[image generated]".to_string(),
-            },
-            metadata_json,
-        },
-    }
-}
-
-/// Clone a WIT StatusUpdate (the generated type doesn't derive Clone).
-fn clone_wit_status_update(update: &wit_channel::StatusUpdate) -> wit_channel::StatusUpdate {
-    wit_channel::StatusUpdate {
-        status: match update.status {
-            wit_channel::StatusType::Thinking => wit_channel::StatusType::Thinking,
-            wit_channel::StatusType::Done => wit_channel::StatusType::Done,
-            wit_channel::StatusType::Interrupted => wit_channel::StatusType::Interrupted,
-            wit_channel::StatusType::ToolStarted => wit_channel::StatusType::ToolStarted,
-            wit_channel::StatusType::ToolCompleted => wit_channel::StatusType::ToolCompleted,
-            wit_channel::StatusType::ToolResult => wit_channel::StatusType::ToolResult,
-            wit_channel::StatusType::ApprovalNeeded => wit_channel::StatusType::ApprovalNeeded,
-            wit_channel::StatusType::Status => wit_channel::StatusType::Status,
-            wit_channel::StatusType::JobStarted => wit_channel::StatusType::JobStarted,
-            wit_channel::StatusType::AuthRequired => wit_channel::StatusType::AuthRequired,
-            wit_channel::StatusType::AuthCompleted => wit_channel::StatusType::AuthCompleted,
-        },
-        message: update.message.clone(),
-        metadata_json: update.metadata_json.clone(),
-    }
-}
-
-/// HTTP response from a WASM channel callback.
-#[derive(Debug, Clone)]
-pub struct HttpResponse {
-    /// HTTP status code.
-    pub status: u16,
-    /// Response headers.
-    pub headers: HashMap<String, String>,
-    /// Response body.
-    pub body: Vec<u8>,
-}
-
-impl HttpResponse {
-    /// Create an OK response.
-    pub fn ok() -> Self {
-        Self {
-            status: 200,
-            headers: HashMap::new(),
-            body: Vec::new(),
-        }
-    }
-
-    /// Create a JSON response.
-    pub fn json(value: serde_json::Value) -> Self {
-        let body = serde_json::to_vec(&value).unwrap_or_default();
-        let mut headers = HashMap::new();
-        headers.insert("Content-Type".to_string(), "application/json".to_string());
-        Self {
-            status: 200,
-            headers,
-            body,
-        }
-    }
-
-    /// Create an error response.
-    pub fn error(status: u16, message: &str) -> Self {
-        Self {
-            status,
-            headers: HashMap::new(),
-            body: message.as_bytes().to_vec(),
-        }
-    }
-}
-
-/// Extract the hostname from a URL string.
-///
-/// Returns `None` for malformed URLs or non-HTTP(S) schemes.
-fn extract_host_from_url(url: &str) -> Option<String> {
-    let parsed = url::Url::parse(url).ok()?;
-    if !matches!(parsed.scheme(), "http" | "https") {
-        return None;
-    }
-    parsed.host_str().map(|h| {
-        h.strip_prefix('[')
-            .and_then(|v| v.strip_suffix(']'))
-            .unwrap_or(h)
-            .to_lowercase()
-    })
-}
-
-/// Pre-resolve host credentials for all HTTP capability mappings.
-///
-/// Called once per callback (in async context, before spawn_blocking) so the
-/// synchronous WASM host function can inject credentials without needing async
-/// access to the secrets store.
-///
-/// Silently skips credentials that can't be resolved (e.g., missing secrets).
-/// The channel will get a 401/403 from the API, which is the expected UX when
-/// auth hasn't been configured yet.
-async fn resolve_channel_host_credentials(
-    capabilities: &ChannelCapabilities,
-    store: Option<&(dyn SecretsStore + Send + Sync)>,
-) -> Vec<ResolvedHostCredential> {
-    let store = match store {
-        Some(s) => s,
-        None => return Vec::new(),
-    };
-
-    let http_cap = match &capabilities.tool_capabilities.http {
-        Some(cap) => cap,
-        None => return Vec::new(),
-    };
-
-    if http_cap.credentials.is_empty() {
-        return Vec::new();
-    }
-
-    let mut resolved = Vec::new();
-
-    for mapping in http_cap.credentials.values() {
-        // Skip UrlPath credentials; they're handled by placeholder substitution
-        if matches!(
-            mapping.location,
-            crate::secrets::CredentialLocation::UrlPath { .. }
-        ) {
-            continue;
-        }
-
-        let secret = match store.get_decrypted("default", &mapping.secret_name).await {
-            Ok(s) => s,
-            Err(e) => {
-                tracing::debug!(
-                    secret_name = %mapping.secret_name,
-                    error = %e,
-                    "Could not resolve credential for WASM channel (auth may not be configured)"
-                );
-                continue;
-            }
-        };
-
-        let mut injected = InjectedCredentials::empty();
-        inject_credential(&mut injected, &mapping.location, &secret);
-
-        if injected.is_empty() {
-            continue;
-        }
-
-        resolved.push(ResolvedHostCredential {
-            host_patterns: mapping.host_patterns.clone(),
-            headers: injected.headers,
-            query_params: injected.query_params,
-            secret_value: secret.expose().to_string(),
-        });
-    }
-
-    if !resolved.is_empty() {
-        tracing::debug!(
-            count = resolved.len(),
-            "Pre-resolved host credentials for WASM channel execution"
-        );
-    }
-
-    resolved
-}
-
-// ============================================================================
-// Attachment Helpers
-// ============================================================================
-
-/// Maximum total attachment size (50 MB).
-const MAX_TOTAL_ATTACHMENT_BYTES: u64 = 50 * 1024 * 1024;
-
-/// Detect MIME type from file extension using the `mime_guess` crate.
-fn mime_from_extension(path: &str) -> String {
-    mime_guess::from_path(path)
-        .first_or_octet_stream()
-        .to_string()
-}
-
-/// Read attachment files from disk and build WIT attachment records.
-///
-/// Validates total size against `MAX_TOTAL_ATTACHMENT_BYTES`.
-fn read_attachments(paths: &[String]) -> Result<Vec<wit_channel::Attachment>, String> {
-    if paths.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    let mut attachments = Vec::with_capacity(paths.len());
-    let mut total_bytes: u64 = 0;
-    let tmp_base = std::path::Path::new("/tmp");
-    let home_base = dirs::home_dir()
-        .map(|h| h.join(".ironclaw"))
-        .unwrap_or_default();
-
-    for path in paths {
-        // Validate paths are under /tmp/ or ~/.ironclaw/ to prevent arbitrary file reads
-        let validated = crate::tools::builtin::path_utils::validate_path(path, Some(tmp_base))
-            .or_else(|_| crate::tools::builtin::path_utils::validate_path(path, Some(&home_base)));
-        let validated = validated.map_err(|e| {
-            format!(
-                "Invalid attachment path '{}': must be under /tmp/ or ~/.ironclaw/: {}",
-                path, e
-            )
-        })?;
-
-        // Pre-check file size before reading into memory to avoid OOM
-        let file_size = std::fs::metadata(&validated)
-            .map_err(|e| format!("Failed to stat attachment '{}': {}", validated.display(), e))?
-            .len();
-        total_bytes += file_size;
-        if total_bytes > MAX_TOTAL_ATTACHMENT_BYTES {
-            return Err(format!(
-                "Total attachment size exceeds {} MB limit",
-                MAX_TOTAL_ATTACHMENT_BYTES / (1024 * 1024)
-            ));
-        }
-
-        let data = std::fs::read(&validated)
-            .map_err(|e| format!("Failed to read attachment '{}': {}", validated.display(), e))?;
-
-        let filename = validated
-            .file_name()
-            .and_then(|n| n.to_str())
-            .unwrap_or("file")
-            .to_string();
-
-        let mime_type = mime_from_extension(path);
-
-        attachments.push(wit_channel::Attachment {
-            filename,
-            mime_type,
-            data,
-        });
-    }
-
-    Ok(attachments)
 }
 
 #[cfg(test)]
